Put `serde` code behind `json` feature flag in `ergotree-ir`
