+++ ergotree-ir/src/chain/votes.rs
@@ -11,55 +11,16 @@ use super::base16_bytes::Base16EncodedBytes;
 #[derive(PartialEq, Eq, Debug, Clone)]
 #[cfg_attr(
     feature = "json",
-    serde(into = "Base16EncodedBytes", try_from = "VotesEncodingVariants")
+    serde(
+        into = "Base16EncodedBytes",
+        try_from = "crate::chain::json::votes::VotesEncodingVariants"
+    )
 )]
 pub struct Votes(pub [u8; 3]);
-
-#[cfg_attr(feature = "json", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Debug, Clone)]
-#[cfg_attr(feature = "json", serde(untagged))]
-#[allow(dead_code)]
-enum VotesEncodingVariants {
-    AsStr(Base16DecodedBytes),
-    /// We need `serde_json::Number` here due to a known `serde_json` bug described here:
-    /// <https://github.com/serde-rs/json/issues/740>. Basically we can't deserialise any integer
-    /// types directly within untagged enums when the `arbitrary_precision` feature is used. The
-    /// workaround is to deserialize as `serde_json::Number` first, then manually convert the type.
-    AsByteArray(Vec<serde_json::Number>), // explorer v1
-}
-
-impl TryFrom<VotesEncodingVariants> for Votes {
-    type Error = VotesError;
-
-    fn try_from(value: VotesEncodingVariants) -> Result<Self, Self::Error> {
-        match value {
-            VotesEncodingVariants::AsStr(bytes) => bytes.try_into(),
-            VotesEncodingVariants::AsByteArray(bytes) => bytes.try_into(),
-        }
-    }
-}
 
 impl From<Votes> for Vec<u8> {
     fn from(v: Votes) -> Self {
         v.0.to_vec()
-    }
-}
-
-impl TryFrom<Vec<serde_json::Number>> for Votes {
-    type Error = VotesError;
-
-    fn try_from(bytes: Vec<serde_json::Number>) -> Result<Self, Self::Error> {
-        let bytes_u8: Vec<u8> = bytes
-            .into_iter()
-            .map(|n| {
-                #[allow(clippy::unwrap_used)]
-                {
-                    n.as_u64().unwrap() as u8
-                }
-            })
-            .collect();
-        let arr: [u8; 3] = bytes_u8.as_slice().try_into()?;
-        Ok(Self(arr))
     }
 }
